Finished project with unit and snapshot tests.

+++ .eslintrc.js
@@ -17,7 +17,7 @@ module.exports = {
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    quotes: [0, 'single']
+    quotes: [0, 'double']
   },
   overrides: [
     {
